add shipwars code
